requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
attrs==23.2.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
Django==5.2.4
django-widget-tweaks>=1.5.0
djangorestframework>=3.14.0
django-simple-history==3.7.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.7.0
geojson==3.1.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
icalendar>=5.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.5.1
murmurhash==1.0.13
mypy==1.17.0
mypy_extensions==1.1.0
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl>=3.1.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==11.2.1
planetary-computer==1.0.0
preshed==3.0.10
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.2.3
pystac==1.10.1
pystac-client==0.8.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qrcode==8.2
referencing==0.35.1
reportlab>=4.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9.1
setuptools==80.9.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
starlette==0.37.2
thinc==8.3.6
tqdm==4.67.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.17.2
yarl==1.20.1
